Refactor: Improve DataDrivenInteractionGenerator
